pyproject.toml
[project]
name = "uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.7.1",
    "google-api-python-client==2.108.0",
    "google-auth-httplib2==0.1.1",
    "google-auth-oauthlib==1.1.0",
    "PyGithub==2.1.1",
    "python-dotenv==1.0.0",
    "pandas==2.1.4",
    "numpy==1.26.2",
    "python-dateutil==2.8.2",
    "fastapi>=0.104.1",
    "uvicorn==0.24.0",
    "pydantic==2.7.2",
    "python-multipart==0.0.9",
]
